feat(chaining): event logic flow
